tests: merge and move specs
